disp: fix compilation issues in msm disp driver
Resolve compilation issues in msm display driver.
Avoid DP compilation as it will be enabled once the
dependent drivers are available.
Disable system cache for pineapple target as the
dependent drivers are not ready yet.

+++ msm/msm_smmu.c
@@ -567,7 +567,7 @@ static int msm_smmu_probe(struct platform_device *pdev)
 	if (!client->dev->dma_parms)
 		client->dev->dma_parms = devm_kzalloc(client->dev,
 				sizeof(*client->dev->dma_parms), GFP_KERNEL);
-	dma_set_max_seg_size(client->dev, DMA_BIT_MASK(32));
+	dma_set_max_seg_size(client->dev, (unsigned int)DMA_BIT_MASK(32));
 	dma_set_seg_boundary(client->dev, (unsigned long)DMA_BIT_MASK(64));
 
 	iommu_set_fault_handler(client->domain,
